requirements/base.txt
Django==1.5.1
Pillow==2.5.1
South==0.7.6
Unipath==1.0
argparse==1.1
bootstrap-admin==0.1.8
django-genericadmin==0.6
django-kronos==0.4
django-model-utils==1.5.0
django-relatedadminwidget==0.0.2
django-resized==0.2.4
django-autocomplete-light
python-dateutil==2.2
six==1.4.1
sqlparse==0.1.11
wsgiref==0.1.2
django-salmonella
